added a few explicit casts to allow compilation on warning-intolerant build systems (which should be ALL build systems!)
Former-commit-id: 8362bb6535467d56874c30a0fa5610fbe88f245e

+++ pwiz/analysis/spectrum_processing/MS2NoiseFilter.cpp
@@ -198,7 +198,7 @@ FilterSpectrum::FilterSpectrum(const MS2NoiseFilter::Config& params_,
         if (params.relaxLowMass && precursorCharge > 0)
         {
             if (min(precursorCharge, (int) ((precursorMZ * precursorCharge) / *lb)) > 1)
-                numMassesInWindow = params.numMassesInWindow * (factor * min(precursorCharge, (int) ((precursorMZ * precursorCharge) / *lb)));
+                numMassesInWindow = (size_t)(params.numMassesInWindow * (factor * min(precursorCharge, (int) ((precursorMZ * precursorCharge) / *lb))));
         }
 
         if (indexValuePairs.size() > numMassesInWindow)
